Split syntax library out from luau in definitions
