+++ lute/luau/CMakeLists.txt
@@ -16,5 +16,5 @@ target_sources(Lute.Luau PRIVATE
 
 target_compile_features(Lute.Luau PUBLIC cxx_std_17)
 target_include_directories(Lute.Luau PUBLIC "include" ${LIBUV_INCLUDE_DIR})
-target_link_libraries(Lute.Luau PRIVATE Lute.Common Lute.Require Lute.Runtime uv_a Luau.Analysis Luau.Ast Luau.Compiler Luau.CLI.lib Luau.Require Luau.VM)
+target_link_libraries(Lute.Luau PRIVATE Lute.Common Lute.Require Lute.Runtime Lute.Syntax uv_a Luau.Analysis Luau.Ast Luau.Compiler Luau.CLI.lib Luau.Require Luau.VM)
 target_compile_options(Lute.Luau PRIVATE ${LUTE_OPTIONS})
