+++ lute/require/src/lutevfs.cpp
@@ -9,6 +9,7 @@
 #include "lute/modulepath.h"
 #include "lute/net.h"
 #include "lute/process.h"
+#include "lute/syntax.h"
 #include "lute/system.h"
 #include "lute/task.h"
 #include "lute/time.h"
@@ -23,16 +24,17 @@ const Luau::DenseHashMap<std::string, lua_CFunction> kLuteModules = []()
     Luau::DenseHashMap<std::string, lua_CFunction> map{""};
     map["@lute/crypto.luau"] = luteopen_crypto;
     map["@lute/fs.luau"] = luteopen_fs;
+    map["@lute/io.luau"] = luteopen_io;
     map["@lute/luau.luau"] = luteopen_luau;
     map["@lute/net/init.luau"] = luteopen_net;
     map["@lute/net/client.luau"] = luteopen_net_client;
     map["@lute/net/server.luau"] = luteopen_net_server;
     map["@lute/process.luau"] = luteopen_process;
-    map["@lute/task.luau"] = luteopen_task;
-    map["@lute/vm.luau"] = luteopen_vm;
+    map["@lute/syntax/cst.luau"] = luteopen_syntax;
     map["@lute/system.luau"] = luteopen_system;
+    map["@lute/task.luau"] = luteopen_task;
     map["@lute/time.luau"] = luteopen_time;
-    map["@lute/io.luau"] = luteopen_io;
+    map["@lute/vm.luau"] = luteopen_vm;
     return map;
 }();
